# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/pub/current/menutree": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "登录管理"
                ],
                "summary": "查询当前用户菜单树",
                "responses": {
                    "200": {
                        "description": "查询结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/schema.ListResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.MenuTree"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/pub/current/password": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "登录管理"
                ],
                "summary": "更新个人密码",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.UpdatePasswordParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.StatusResult"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/pub/current/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "登录管理"
                ],
                "summary": "获取当前用户信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.UserLoginInfo"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/pub/login": {
            "post": {
                "tags": [
                    "登录管理"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.LoginParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.LoginTokenInfo"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/pub/login/captcha": {
            "get": {
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "登录管理"
                ],
                "summary": "响应图形验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "验证码ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "重新加载",
                        "name": "reload",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "图形验证码"
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
...
        "/api/v1/pub/refresh-token": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "登录管理"
                ],
                "summary": "刷新令牌",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.LoginTokenInfo"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.ErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
        "schema.ErrorResult": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误项",
                    "$ref": "#/definitions/schema.ErrorItem"
                }
            }
        },
...
        "schema.ListResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "object"
                },
                "pagination": {
                    "$ref": "#/definitions/schema.PaginationResult"
                }
            }
        },
...
        "schema.LoginParam": {
            "type": "object",
            "required": [
                "captcha_code",
                "captcha_id",
                "password",
                "user_name"
            ],
            "properties": {
                "captcha_code": {
                    "description": "验证码",
                    "type": "string"
                },
                "captcha_id": {
                    "description": "验证码ID",
                    "type": "string"
                },
                "password": {
                    "description": "密码(md5加密)",
                    "type": "string"
                },
                "user_name": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "schema.LoginTokenInfo": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "访问令牌",
                    "type": "string"
                },
                "expires_at": {
                    "description": "令牌到期时间戳",
                    "type": "integer"
                },
                "token_type": {
                    "description": "令牌类型",
                    "type": "string"
                }
            }
        },
...
        "schema.MenuAction": {
            "type": "object",
            "required": [
                "code",
                "menu_id",
                "name"
            ],
            "properties": {
                "code": {
                    "description": "动作编号",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "menu_id": {
                    "description": "菜单ID",
                    "type": "string"
                },
                "name": {
                    "description": "动作名称",
                    "type": "string"
                },
                "resources": {
                    "description": "资源列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.MenuActionResource"
                    }
                }
            }
        },
        "schema.MenuActionResource": {
            "type": "object",
            "required": [
                "method",
                "path"
            ],
            "properties": {
                "action_id": {
                    "description": "菜单动作ID",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "method": {
                    "description": "资源请求方式(支持正则)",
                    "type": "string"
                },
                "path": {
                    "description": "资源请求路径（支持/:id匹配）",
                    "type": "string"
                }
            }
        },
        "schema.MenuTree": {
            "type": "object",
            "properties": {
                "actions": {
                    "description": "动作列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.MenuAction"
                    }
                },
                "children": {
                    "description": "子级树",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.MenuTree"
                    }
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级ID",
                    "type": "string"
                },
                "parent_path": {
                    "description": "父级路径",
                    "type": "string"
                },
                "router": {
                    "description": "访问路由",
                    "type": "string"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "show_status": {
                    "description": "显示状态(1:显示 2:隐藏)",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                }
            }
        },
        "schema.PaginationResult": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "schema.Role": {
            "type": "object",
            "required": [
                "name",
                "role_menus",
                "status"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "role_menus": {
                    "description": "角色菜单列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.RoleMenu"
                    }
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "schema.RoleMenu": {
            "type": "object",
            "required": [
                "action_id",
                "menu_id",
                "role_id"
            ],
            "properties": {
                "action_id": {
                    "description": "动作ID",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "menu_id": {
                    "description": "菜单ID",
                    "type": "string"
                },
                "role_id": {
                    "description": "角色ID",
                    "type": "string"
                }
            }
        },
        "schema.StatusResult": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "状态(OK)",
                    "type": "string"
                }
            }
        },
        "schema.UpdatePasswordParam": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "新密码(md5加密)",
                    "type": "string"
                },
                "old_password": {
                    "description": "旧密码(md5加密)",
                    "type": "string"
                }
            }
        },
...
        "schema.UserLoginInfo": {
            "type": "object",
            "properties": {
                "real_name": {
                    "description": "真实姓名",
                    "type": "string"
                },
                "roles": {
                    "description": "角色列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Role"
                    }
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "string"
                },
                "user_name": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
    }
}